DCS-863 Restructuring routes
