+++ server/routes/creatingReports/changePrison.ts
@@ -1,7 +1,7 @@
 import { Request, Response } from 'express'
-import LocationService from '../services/locationService'
-import DraftReportService from '../services/drafts/draftReportService'
-import { SystemToken } from '../types/uof'
+import LocationService from '../../services/locationService'
+import DraftReportService from '../../services/drafts/draftReportService'
+import { SystemToken } from '../../types/uof'
 
 export default class ChangePrisonRoutes {
   constructor(
