+++ server/routes/creatingReports/createReport.test.ts
@@ -1,8 +1,8 @@
 import request from 'supertest'
-import DraftReportService from '../services/drafts/draftReportService'
-import { appWithAllRoutes, user } from './testutils/appSetup'
+import DraftReportService from '../../services/drafts/draftReportService'
+import { appWithAllRoutes, user } from '../__test/appSetup'
 
-jest.mock('../services/drafts/draftReportService')
+jest.mock('../../services/drafts/draftReportService')
 
 const draftReportService = new DraftReportService(null, null, null, null, null, null) as jest.Mocked<DraftReportService>
