+++ server/routes/creatingReports/incidentDetails.test.ts
@@ -1,14 +1,14 @@
 import request from 'supertest'
-import { appWithAllRoutes, user } from './testutils/appSetup'
-import { toDate } from '../utils/dateSanitiser'
-import DraftReportService from '../services/drafts/draftReportService'
-import OffenderService from '../services/offenderService'
-import LocationService from '../services/locationService'
-import type { Prison } from '../data/elite2ClientBuilderTypes'
+import { appWithAllRoutes, user } from '../__test/appSetup'
+import { toDate } from '../../utils/dateSanitiser'
+import DraftReportService from '../../services/drafts/draftReportService'
+import OffenderService from '../../services/offenderService'
+import LocationService from '../../services/locationService'
+import type { Prison } from '../../data/elite2ClientBuilderTypes'
 
-jest.mock('../services/drafts/draftReportService')
-jest.mock('../services/offenderService')
-jest.mock('../services/locationService')
+jest.mock('../../services/drafts/draftReportService')
+jest.mock('../../services/offenderService')
+jest.mock('../../services/locationService')
 
 const draftReportService = new DraftReportService(null, null, null, null, null, null) as jest.Mocked<DraftReportService>
 const offenderService = new OffenderService(null) as jest.Mocked<OffenderService>
