+++ server/routes/creatingReports/incidentDetails.ts
@@ -1,15 +1,15 @@
 import moment from 'moment'
 import { Request, Response } from 'express'
-import { isNilOrEmpty, firstItem } from '../utils/utils'
-import types from '../config/types'
-import { processInput } from '../services/validation'
-import { nextPaths, full, partial } from '../config/incident'
-import type OffenderService from '../services/offenderService'
-import type LocationService from '../services/locationService'
-import type DraftReportService from '../services/drafts/draftReportService'
-import type { ParsedDate } from '../utils/dateSanitiser'
-import type { SystemToken } from '../types/uof'
-import { isReportComplete } from '../services/drafts/reportStatusChecker'
+import { isNilOrEmpty, firstItem } from '../../utils/utils'
+import types from '../../config/types'
+import { processInput } from '../../services/validation'
+import { nextPaths, full, partial } from '../../config/incident'
+import type OffenderService from '../../services/offenderService'
+import type LocationService from '../../services/locationService'
+import type DraftReportService from '../../services/drafts/draftReportService'
+import type { ParsedDate } from '../../utils/dateSanitiser'
+import type { SystemToken } from '../../types/uof'
+import { isReportComplete } from '../../services/drafts/reportStatusChecker'
 
 const formName = 'incidentDetails'
