+++ server/routes/creatingReports/reportUseOfForce.test.ts
@@ -1,9 +1,9 @@
 import request from 'supertest'
-import { DraftReportService, OffenderService } from '../services'
-import { appWithAllRoutes } from './testutils/appSetup'
+import { DraftReportService, OffenderService } from '../../services'
+import { appWithAllRoutes } from '../__test/appSetup'
 
-jest.mock('../services/offenderService')
-jest.mock('../services/drafts/draftReportService')
+jest.mock('../../services/offenderService')
+jest.mock('../../services/drafts/draftReportService')
 
 const offenderService = new OffenderService(null) as jest.Mocked<OffenderService>
 const draftReportService = new DraftReportService(null, null, null, null, null, null) as jest.Mocked<DraftReportService>
